requirements.txt
streamlit>=1.25.0
requests>=2.28.0
python-dotenv>=1.0.0
google-genai>=0.4.0
sentence-transformers>=2.2.2
chromadb>=0.3.29
ujson>=5.8.0
pytz
python-dateutil
